react-router-dom installed and refactored packages (page-title and residences-list)
