+++ package.json
@@ -7,6 +7,7 @@
     "react": "16.9.0",
     "react-bootstrap": "1.0.0-beta.12",
     "react-dom": "16.9.0",
+    "react-router-dom": "5.1.2",
     "react-scripts": "3.1.1",
     "styled-components": "4.4.0",
     "typedi": "0.8.0"
