+++ src/app.tsx
@@ -1,14 +1,15 @@
 import React from 'react';
-import { Dashboard } from './app/dashboard/dashboard.component';
-import { PageTitle } from './components/mol.page_title';
+import { SearchForm } from './app/search-form/search-form.component';
+import { PageTitle } from './components/mol.page-title';
+import Container from 'react-bootstrap/Container';
 
 export class TravelSimpleApp extends React.Component {
 	render() {
 		return (
-			<>
+			<Container fluid={true}>
 				<PageTitle />
-				<Dashboard />
-			</>
+				<SearchForm />
+			</Container>
 		);
 	}
 }
