--- src/app/dashboard/index.ts
@@ -1,2 +1,0 @@
-export * from './dashboard.component';
-export * from './residence-list-view.component';
