+++ src/app/index.ts
@@ -1,1 +1,1 @@
-export * from './dashboard';
+export * from './search-form';
