+++ src/app/search-form/index.ts
@@ -1,0 +1,1 @@
+export * from './search-form.component';
