--- src/components/atm.headers/headers.style.tsx
@@ -1,12 +1,0 @@
-import styled from "styled-components";
-import { FontSize, DefaultColor } from "../../resources";
-
-export const H1 = styled.h1`
-  font-size: ${FontSize.Large};
-  color: ${DefaultColor.Black};
-`;
-
-export const H3 = styled.h3`
-  font-size: ${FontSize.Medium};
-  color: ${DefaultColor.Black};
-`;
