--- src/components/atm.headers/index.ts
@@ -1,1 +1,0 @@
-export * from './headers.style';
