+++ src/components/index.ts
@@ -1,2 +1,4 @@
 export * from './typography.style';
 export * from './mol.residence-view';
+export * from './mol.page-title';
+export * from './mol.residences-list/residences-list.component';
