+++ src/components/mol.residences-list/residences-list.component.style.tsx
